refactor: share stripe billing customer lookup
